+++ src/AddTodoForm.jsx
@@ -1,0 +1,12 @@
+function AddTodoForm() {
+
+    return (
+        <form>
+            <label htmlFor="todoTitle">Title</label>
+            <input id="todoTitle"></input>
+            <button type="submit">Add</button>
+        </form>
+    )
+}
+
+export default AddTodoForm;
